refactor internal indexeddb query builder
